requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
twilio==9.4.0
SQLAlchemy==2.0.25
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.4
pytest-flask==1.3.0
